updated backend process for file deletion in regards to images

+++ Views/Pages/Backend/AdminCp/Images/manage.blade.php
@@ -12,20 +12,24 @@
         </div>
     </div>
     <div class="container">
-        
+        @if(($image->count() == 1))
+            <div class="row">
+                <div class="col-sm-12 head">Manage Image {{$image->title}}  </div>
+
+                <img src="/img/uploads/{{$image->name}}" height="200" width="200"  class="m-2" alt="{{$image->title}}">
+
+                {{--            Add Ability for admins to delete and unlink images here.--}}
+
+                <div class="col-sm-12 text-center"><a href="{{$url->make("auth.admin.images.delete",["id"=>base64_encode($image->id)])}}">Delete Image</a></div>
+            </div>
         {{--    This section needs to be linkjed to the users first grouped together and add a count section--}}
-        <div class="row">
-            <div class="col-sm-12 head">Manage Image {{$image->title}}  </div>
-
-            <img src="/img/uploads/{{$image->image_name}}" height="200" width="200" alt="{{$image->title}}">
-
-{{--            Add Ability for admins to delete and unlink images here.--}}
-
-            <form action="# class="tld-form">
-
-            </form>
-        </div>
+        @else
+            <div class="row">
+                <div class="col-sm-12"><a href="{{$url->make("auth.admin.images.home")}}">No Image found Return to image list</a></div>
+            </div>
+    @endif
     </div>
+
 
 
 @endsection
